+++ non_catalog_apps/flizzer_tracker/docs/changelog.md
@@ -1,19 +1,27 @@
-# Flizzer Tracker v0.4 #
+# Flizzer Tracker v0.10 #
+## Fixed ##
+- App not responding to keypresses
+
+# Flizzer Tracker v0.9 #
 
 ## Added ##
 - Vxx effect (detune)
 
-# Flizzer Tracker v0.3 #
+# Flizzer Tracker v0.8 #
 
 ## Added ##
 - Qxx effect (set tracker engine rate)
 
-# Flizzer Tracker v0.2 #
+# Flizzer Tracker v0.7 #
 
 ## Added ##
 - Save/load instruments in separate .fzi files
 - Pattern editor now occupies full screen when you focus on it
 - Copypaste menu (hold Back to open it when focused on pattern editor), operates on whole patterns
+
+# Flizzer Tracker v0.2-v0.6 #
+
+- Small fixes for new firmware versions
 
 # Flizzer Tracker v0.1 #